requirements.txt
asgiref==3.8.1
astroid==3.3.8
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coverage==7.6.10
dill==0.3.9
Django==4.2.16
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.10.1
pylint==3.3.3
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.2
tomlkit==0.13.2
tzdata==2024.2
urllib3==2.2.3
